Move cast impls into kernels::cast
