+++ candle-metal-kernels/src/lib.rs
@@ -8,10 +8,11 @@ pub mod utils;
 pub use err::MetalKernelError;
 pub use kernel::Kernels;
 pub use kernels::{
-    call_arg_sort, call_binary_contiguous, call_binary_strided, call_const_set_contiguous,
-    call_const_set_contiguous_tiled, call_const_set_strided, call_mlx_arg_sort, call_mlx_gemm,
-    call_quantized_matmul_mm_t, call_quantized_matmul_mv_t, call_unary_contiguous,
-    call_unary_contiguous_tiled, call_unary_strided, unary, GemmDType, GgmlDType,
+    call_arg_sort, call_binary_contiguous, call_binary_strided, call_cast_contiguous,
+    call_cast_strided, call_const_set_contiguous, call_const_set_contiguous_tiled,
+    call_const_set_strided, call_mlx_arg_sort, call_mlx_gemm, call_quantized_matmul_mm_t,
+    call_quantized_matmul_mv_t, call_unary_contiguous, call_unary_contiguous_tiled,
+    call_unary_strided, unary, GemmDType, GgmlDType,
 };
 use metal::{
     BlitCommandEncoder, Buffer, CommandQueue, ComputeCommandEncoder, ComputePipeline,
@@ -95,63 +96,6 @@ pub fn call_copy2d(
     encoder.use_resource(input, MTLResourceUsage::Read);
     encoder.use_resource(output, MTLResourceUsage::Write);
     encoder.dispatch_threads(grid_dims, group_dims);
-    Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn call_cast_contiguous(
-    device: &Device,
-    ep: impl EncoderProvider,
-    kernels: &Kernels,
-    kernel_name: &'static str,
-    length: usize,
-    input: BufferOffset,
-    output: &Buffer,
-) -> Result<(), MetalKernelError> {
-    let pipeline = kernels.load_pipeline(device, Source::Cast, kernel_name)?;
-
-    let encoder = ep.encoder();
-    let encoder: &ComputeCommandEncoder = encoder.as_ref();
-    encoder.set_compute_pipeline_state(&pipeline);
-
-    set_params!(encoder, (length, &input, output));
-
-    let (thread_group_count, thread_group_size) = linear_split(&pipeline, length);
-    encoder.use_resource(input.buffer, MTLResourceUsage::Read);
-    encoder.use_resource(output, MTLResourceUsage::Write);
-    encoder.dispatch_thread_groups(thread_group_count, thread_group_size);
-    Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn call_cast_strided(
-    device: &Device,
-    ep: impl EncoderProvider,
-    kernels: &Kernels,
-    kernel_name: &'static str,
-    shape: &[usize],
-    input: BufferOffset,
-    input_strides: &[usize],
-    output: &Buffer,
-) -> Result<(), MetalKernelError> {
-    let pipeline = kernels.load_pipeline(device, Source::Cast, kernel_name)?;
-
-    let encoder = ep.encoder();
-    let encoder: &ComputeCommandEncoder = encoder.as_ref();
-    encoder.set_compute_pipeline_state(&pipeline);
-
-    let length: usize = shape.iter().product();
-
-    set_params!(
-        encoder,
-        (length, shape.len(), shape, input_strides, &input, output)
-    );
-
-    let (thread_group_count, thread_group_size) = linear_split(&pipeline, length);
-
-    encoder.use_resource(input.buffer, MTLResourceUsage::Read);
-    encoder.use_resource(output, MTLResourceUsage::Write);
-    encoder.dispatch_thread_groups(thread_group_count, thread_group_size);
     Ok(())
 }
